INC-50: Ability to select current location (aka "wing") (#8)
* INC-50: Ability to select current location/wing

* Guard against stale location on Case Load change

Check whether the currently active location/wing belongs to the currently
active case load. This is to invalidate the active location when a user
changed their active load.

* Tests for change location functionality

* Mechanism to control/check on session during tests

* Integration tests for location selection

- Fix existing integration tests
- Mock Prison API calls
- Renamed `IndexPage` to `BehaviourEntriesPage`

* Git ignores macOS' .DS_Store files

* Use destructured user variable

* Added Prison API to health checks

* Removed `/api` from Prison API URL config

The trailing `/api/` is present in almost all the endpoint but not in the
health check endpoint.

As I've added a check for the Prison API, the trailing `/api/` would
make the check fail, so I've updated the config to not have the `/api`
prefix anymore.