BehaviourService test and test data mechanism
Added a basic mechanism to get some test data. This is reduces duplication
or unnecessary level of details in tests but at the same time is still
flexible enough to have control over the test data.

Test data functions are in modules under `/server/testData/` and are name
`getTest${TYPE}` to reduce ambiguity.

Test data functions have good defaults with all fields options so that
values of a specific type can be obtained easily.