Increased Incentives API timeouts to 60s (#10)
* Increased Incentives API timeouts to 60s

Some requests, expecially while API cache is not warm, can take a bit long
so we're increasing the timeouts to try and reduce/avoid them.

* Jest: Added --detectOpenHandles --restoreMocks

For some reason the tests started to hang in CircleCI (tests are passing
locally). This could help give us some clues or potentially even solve
the problem.