Simplified location selection
The location is not stored in the session anymore (there is no concept
of "active" location). This is unnecessary really given the location
prefix is part of the URL for the behaviour entries page now.

This also means that there is not need to check that the "active" location
is part of the active case load (in case a user changed their active case
load in DPS and the old "active" location was now invalid). Therefore
the 'checkActiveLocation' middleware is also gone.