INC-153: Incentive information page has RESTful URL (#20)
* Link(s) in header takes to DPS home page.

* Renamed 'index' routes module to 'incentivesTable'

This has no other impact just yet but it makes it clear this is the
routing related to the incentive information table.
This is currently the home page but navigation is subject to change.

* Copied 'Card' component from DPS

* Layout tweaks

- added breadcrumbs
- removed 'Back' link in location selection page
- removed 'Cancel' button in location selection page
- removed padding before active case load name

* Added homepage with link to table in a card

* Incentives table has REST-ful and reproducible URL

This table is not shown in the homepage anymore but it has its own
URL which includes the location prefix, e.g. `/incentive-summary/MDI-2`.

* Changed location select route

Renaming it to just `/select-location` for simplicity (and it will not
technically change anything in the future as the location will not be
stored in the session)

* Simplified location selection

The location is not stored in the session anymore (there is no concept
of "active" location). This is unnecessary really given the location
prefix is part of the URL for the behaviour entries page now.

This also means that there is not need to check that the "active" location
is part of the active case load (in case a user changed their active case
load in DPS and the old "active" location was now invalid). Therefore
the 'checkActiveLocation' middleware is also gone.

* Fix integration tests

* Updated copy to be closer to latest version

- header for "behaviour entries" page is now 'incentive information'
- added 'Incentive levels' sub-header above levels summary
- updated table headers to 'Positive behaviours'/'Negative behaviours'
  (was 'All positive behaviours'/'All negative behaviours')
- also, increased size of location select

* Header now shows "HMPPS Digital Prison Services"

Consistent to other DPS services and how the prototype looks.