INC-162: unnecessary API calls when getting prisoner images
The prisoner images route was initially using the `standardRouter` which
was populating the current user/caseloads making additional requests.

This is useful when showing a page but a waste when a browser is getting
prisoner images' raw data.

That route now uses a simpler `imageRouter` which doesn't use the
`populateUser` middleware...however, these API calls were still happening.

The reason for this is order of the routes wiring. The `/` route (which
still uses the `standardRouter`) was still running because it was more
generic. Move it last means these unnecessary requests are not made anymore.

NOTE: I've added a regression test - which would fail if the
`app.use('/', ...` is moved back above the prisoner images route) -
and also a test for the authentication requirement of this endpoint.