INC-212: Zebra stripes for table and other minor UI tweaks (#26)
* Converted stylesheets to SCSS from indented syntax

The SCSS syntax is newer and more similar to vanilla CSS and therefore
a bit more natural.

* Tweaks to incentives table style

- use zebra striping for rows as per design
- vertically center cells text as per design

* Moved wing description down, next to 'Select another location'

As per latest version of Miro board. This is to improve the "wings"
naming situation as the wing description could be anything from
one letter "B" to "B-wing" or even something completely different like
"Houseblock 22".