INC-224: track Incentive information page events in GA (#29)
* Middleware to generate CSP nonce

See Content Security Policy docs:

https://content-security-policy.com/nonce/

* Added .well-known/security.txt

as per MoJ Security guidance:

https://security-guidance.service.justice.gov.uk/implement-security-txt/#implementing-securitytxt

* Added Google Analytics to layout

also addes `analytics.googleAnalyticsId`/`GOOGLE_ANALYTICS_ID` config
and updated Content Security Policy settings so that GA works.

See:
- https://developers.google.com/analytics/devguides/collection/analyticsjs
- https://insightwhale.com/how-to-use-a-content-security-policy-with-google-analytics/

* Corrected Incentive information page title

* Renamed application name variable to 'HMPPS Incentives'

removed the 'UI' suffix as it's a technicality and this shouldn't be
appearing in the pages title (which is also used in Google Analytics)

* GA: Send events on incentive table's links clicks

* GA: Send events on incentive table's tabs clicks

Send Google Analytics events for when a tab gets 'focus'. This is better
than binding the 'click' event because it happens when clicked but also
when using keyboard shortcuts.

Also made event selectors more specific to avoid attaching event handlers
twice.

* GA: Send events on incentive table sorting

When a user sorts the incentive table, an event is sent to GA

In Google Analytics these events would look like this:

('Incentive information > Sorted table', 'by days on level (descending)', ${PRISON_ID})

* Read `GOOGLE_ANALYTICS_ID` config from k8s secret

* GA: Check if available also on client side

This accounts for potential timing issues causing `ga()` function to
not be available and it makes this easier to extract in separate file
later on.