Added phase tag (dev/pre-prod) consistent to DPS
Similarly to what already happens in DPS, on dev/preprod a tag is
shown to make it a bit more evident where you are.

On production the tag is not shown (again like in DPS).

Also removed the margin-top in line with DPS style of caseload selection.

NOTE: The environment variable is called `PHASE_NAME` for consistency
unlike in DPS where this is called 'SYSTEM_PHASE' for some reason.