INC-249: Middleware refactoring
Removed duplication between real express app and test express app.

The test express app should be as similar as possible to the real expess
app (the one determining how users' traffic is handled). This should
reduce chances of tests running against a different version of the routing
as before.

In addition, now to add a route only one file needs to change instead of
two (easy to miss)