Incentive summary Sass: show tabs headings on mobile
We want to hide those on larger screens as the incentive level is already
prominent in the tabs.

However on mobile the tabs are shown differently and we want to continue
to see the tabs heading so that the user doesn't confused regarding the
incentive level list their looking at.