Removed (unused) jQuery UI and simplified CSP
jQuery UI was included in all pages (as it was part of the layout
template) but it doesn't seem to be used anywhere.

Even worse, the js/css wasn't served by the application itself which is
a potential security issue.

Remove these means we can remove the CSP rules to allow requests to
code.jquery.com. It also means we can remove the
`crossOriginEmbedderPolicy: true` option which was added to avoid the
related CSP error.