Redis: Attach error handler on client creation
This was only done in the `TokenStore` class but the Redis client is
created and consumed also by the Express session.

This is important because unhandled error would crash the entire node
application. After this change when there are Redis problems these would
be logged but the application would continue to be up - albeit not
functioning - until these problems are resolved (e.g. Redis connection to
Redis gets re-enstablished).