Reuse Redis client in Incentive information request handler
Instead of creating a new Redis client/TokenStore/HmppsAuthClient on each
request, create this once and reuse across all the requests.

These shouldn't have state (1) and creating a new Redis client/connection
for each new request may be the cause or one of the causes of the Redis
crash we're seeing. Maybe this was somehow causing a resource exhaustion
problem (can't confirm).

This should be more sensible than creating a new HmppsAuthClient instance
for each new request.

(1) The HmppsAuthClient is simply used to make HTTP requests to get a
system token for the user.
The Token store is used to cache these in Redis but the object/instance
itself doesn't have state (aside from the Redis client).
The Redis client also doesn't have any instance aside from the connection
to Redis.
Conclusion: Redis itself is maintaining the state therefore it should be
safe to reuse the same HmppsAuthClient instance across for handling all
requests.