Redis: on 'error', show which Redis client it was
Currently, we instanciate 3 Redis clients. We're seeing some errors
but it's difficult to tell which of the clients is affected.

A string identifying the redis client is now required when calling
`createRedisClient()` and this is shown when an error occurred.

This should help distinguish between then and give us a better idea of
the order of things.