Added /throw-test-error endpoint for debugging purposes
Added to be able to manually trigger errors and see what happens
in a production-like environment (e.g. does Sentry get notified? or the
error is simply logged?)

By default the endpoint throw a Redis' `The client is closed` error
syncronously. Optionally it's possible to pass the `async=true` query
parameter to throw an error asyncronously (thrown in a setTimeout callback)

NOTE: This is behind a feature flag so it would not be available unless the
`FEATURE_ADD_TEST_ERROR_ENDPOINT` setting is set.