Adjust incentives table headers size depending on columns
We reduced the font size of the headers because we had an high number
of columns/text in the headers.

We're now hiding the 'Days since last review'/'Days on level' columns
(behind feature flag) so we're restoring the default table headers style

In practice this means that when we're showing the columns
(`featureFlags.hideDaysColumnsInIncentivesTable=='false'`)
we're adding a CSS class to reduce the headers font size. When we're hiding
these columns we don't add this CSS class and use GDS component's default
style.