Remove "Days" columns from markup
Before this commit the 'Days on level'/'Days since last review' columns
were simply hidden using a CSS class when the 'hideDaysColumnsInIncentivesTable'
was set. Client would still receive the data.

This change adds some nunjuck logic which means client will only get
the columns data if 'hideDaysColumnsInIncentivesTable' flag is not set.

This sidestep the failing integration tests on CI (which pass locally).
Tests relied on selecting only visible columns but for some reason this
didn't work in the same manner as locally in CI.