INC-443: Simplified nunjucks templates using `rejectattr`
In two templates we were showing certain elements in the page only if
a given feature flag was on or off.

We were using `Array.prototype.push()` inside an if statement.

This works but it's a bit hairy and it can be simplified by using
nunjucks' builtin `rejectattr`:

https://mozilla.github.io/nunjucks/templating.html#rejectattr-only-the-single-argument-form

This filters out elements of an array if they have a given property
set to `true`.