Cache Analytics' stitched tables values
In-memory cache for Analytics' stiched tables.

The stitched tables are cached based on the source S3 table name
and the columns selected. Adding more graphs using the same data but
different columns should not cause problems in the future.

This cache will be shared across all the requests served from the same
Node process. Size of biggest stitched table is currently ~30MB which
is small enough to be kept around the process, this means we avoid the
overhead of talking to Redis.
Further clear this cache would be as simple as restart the k8s pods (if
we need to do so).

NOTE: We only cache the matrix result with the filtered columns, the
request handlers will still do the filtering by prison and other aggregations.
This is by design because, the filtering/aggregation step is relatively
quick and to cache those very specific/small artefacts would probably mean
a lot of cache misses (as cache would be only useful on a prison-by-prison
basis).