AnalyticsService: refactored findTable()
Extracted logic to find latest table data (and date/modified metadata)
and the reading/parsing of the table itself.

Most of the client code at this point still uses `findTable()` which
combines these 2 things, that method is behaving as before.

This is preparation work for potentially updating the caching mechanism
to always check if fresh data is available on S3.