StitchedTablesCache: Use S3 modified to determine when cache is stale
Use modified datetime from S3 to determine whether the cached stitched
table is fresh instead of a time-based expiration mechanism.

The cache now checks the modified from S3 and compares it with the modified
in the cache. If these don't match it concludes cache is stale and gets
fresh data from S3.

The drawback of this approach is that cache always list objects in S3
to check if cache is fresh. The positive is that when data in S3 changes
the cache detects it and updates itself.

NB: Cache considers any different modified as a sign the cache is stale -
even older modified. This is by design and means potentially if a newer
S3 object is deleted for whatever reason (e.g. file had incorrect data)
the cache will still return the latest value in S3 instead of sticking
to newer but potentially wrong cached data.