INC-544: Moved/split 'Analytics' integration tests in 3 files
These tests are related (hence in the same directory) but fairly independent
(so they can go in different files).

We usually work on different pages so it's nice to be able to run tests
for a specific page on their own without having to comment tests out.
Plus the file was getting quite big so three files of 100-200 lines are better
than one huge file (files are still on the big side but it's an improvement)