INC-572: other PC trends dropdowns don't reset on submit
In the Protected Characteristic page there are 2 charts with 2 separate
dropdown to select the PC group to display data for. These 2 dropdowns are
independent which means we have 2 separate query params for them.

However the 2 corresponding forms need to re-submit the value for the
other chart PC group or that would be reset.