INC-572: Use "dropdown" consistently instead of "select"
Dropdown has been used in the team as it's "higher level" and people seems
to get confused with the word "select" because it's usually a verb (and
not everyone is familiar with HTML's `<select>` element).

PS: GovUK Design System still uses the word "select", in fact there is a
"Select" component. I'm not changing things like HTML IDs/classes for this
reason.