INC-572: Dropdown for Incentive levels PC trends (#194)
* INC-572: Use correct hyphen character for charts content

instead of usual `-`.

* INC-572: Drop down for Incentive levels PC trends

Added a macro as dropdown markup is very similar to other dropdown in page.

* INC-572: other PC trends dropdowns don't reset on submit

In the Protected Characteristic page there are 2 charts with 2 separate
dropdown to select the PC group to display data for. These 2 dropdowns are
independent which means we have 2 separate query params for them.

However the 2 corresponding forms need to re-submit the value for the
other chart PC group or that would be reset.

* INC-572: Use "dropdown" consistently instead of "select"

Dropdown has been used in the team as it's "higher level" and people seems
to get confused with the word "select" because it's usually a verb (and
not everyone is familiar with HTML's `<select>` element).

PS: GovUK Design System still uses the word "select", in fact there is a
"Select" component. I'm not changing things like HTML IDs/classes for this
reason.