INC-596: Updated truncate table script for new behaviour entries source tables
For National/Regional level behaviour entries data is taken from 2 new
source tables.

NOTE: For National level the source table doesn't have the prison column
so it's not possible to filter by prison. Also it doesn't make much sense
as it's more useful to have the various regions in the test data.