INC-606: Simplified PGD Region service/mapping
We don't need to hard-code the list of prisons in each group (not for now
at least) as this is somehow implicit in the source data.

This means we can keep the mapping from PGD region code to PGD region name
is a simple object.

NOTE: For now I've retained the existing public interface of this small
class but if we think some of this logic is simple enough we could have
the client code use the mapping constant directly.