INC-627: Updated PGD region names and updated test data (#236)
* INC-627: Added most recent PGD Region names to service

Existing names were tweaked and list is now "final".

I added a test that reads the national behaviour entries table and checks
that all PGD region names in there are recognised by the `PgdRegionService`
class. This should help quickly identify changes to this list when the
test data will be updated.

NOTE: This test currently fails as I haven't updated the tests files just
yet.

* Fix typo in mockSdkS3ClientResponse() method name

* INC-627: Added new PGD region names

- added "final" mapping of PGD regions
- PGD region names were tweaked
- re-generated test test data
  - to get latest PGD region names
  - also added new 'prison_name' column

* Husky: Small test to check test data doesn't contain Booking IDs

They're sentive and we shouldn't commit them.

NOTE: This test is not gonna cover all possible ways of accidentally
commit sensitve data but it's a start and it's better than nothing.

* INC-627: Updated unit tests numbers to match updated source files

* INC-627: Updated int. tests to match updated test source data

* INC-627: Added `prison_name` to AnalyticsServiceTypes

* INC-627: Moved check for PGD region names to truncate script

This check was previously part of `PgdRegionService` tests and was only
checking one of the files.

This check is now done in the `truncate-source-table` script before writing
the test data file and it's checking the PGD Region names in all the source
files.