INC-632: Refactoring of analytics request handlers and surrounding code (#248)
* Extracted the national/regional logic out of requests handlers

NOTE: This is a first stab at it, logic is mostly as before but we may
simplify/change it.

* INC-632: AnalyticsView don't throw exception

Added a method to determine whether the provided PGD region is valid
or not. This is used to redirect the user in the request handler.

* INC-632: removed AnalyticsService types, methods use view

Removed types, constants and logic related to determine whether view level
was national, regional, or by prison.

`AnalyticsService` methods now get an `AnalyticsView` object which they
can use both to get the required filtering and also to determine the view
level (using `view.isNational()`, etc...)

* INC-632: keep viewType in AnalyticsView state

This is preparation work for simplifing request handler even more by
avoiding having to set URL function there and pass to all `analyticsService`
methods.

* INC-632: Pass AnalyticsView to service to determine URL functions

By passing the view to the `AnalyticsService` constructor the request
handler don't need to pass this in anymore.

NOTE: Next step will be to remove the `AnalyticsView` parameter from the
service methods - now that this is part of the instance state we don't
need to pass it in anymore.

* INC-632: AnalyticsService methods don't need view parameter

The `AnalyticsService` now holds the view in its state, so we don't need
to pass it every time as an argument.

* INC-632: AnalyticsService: href for 'All' is now `null` for consistency

The `UrlForLocationFunction` function type returns a string with the URL
or `null`. In some charts (e.g. regional) we may not want to have links
and the `href` property in these cases would be `null`.

Updating `href` property for 'All' columns to also be `null` for consistency.

* INC-632: Moved `linkTo()` logic into `AnalyticsView`

It's simple enough, it's based on information the view already possess
and we already pass the view to the templates for rendering so it makes
sense to incorporate this.

* INC-632: Moved `AnalyticsView` into services/

This is a service object more than it is a "route"/request handler.

* INC-632: Linting

* INC-632: Added tests for AnalyticsView class

* INC-632: Improved tests for AnalyticsView.isValidPgdRegion()

as suggested in code review