INC-626: Heatmap All row calculated from different source table (#251)
* INC-626: Renamed TableType.behaviourEntriesPrison for consistency

* INC-626: Source table for national behaviour entries 'All'

'All' figures for National level will come from yet another source table
(`behaviours_28d_national`)

* INC-626: Way to get source table for 'All' row

The 'All' row for the behaviour entries heatmap will need to be calculated
from data coming from a separate (level up) source table.

e.g. for prison level, 'All' row will be calculated based on PGD region
level.

* INC-626: Updated analytics row types

To release the constraint that first value of row is a `string`.

This is because in the case of calculations for the `All` row for national
level rows will simply have `['positives', 'negatives']` which are of
type `[number, number]`.

* INC-626: more useful error for getPrisonersWithEntriesByLocation()

As this method could use one of few source tables, depending on level,
it's useful to include this as part of the exception.
This will help if/when we'll need to investigate any data problem.

* INC=626: Heatmap chart 'All' row calculated from level above's data

The behaviour entries Heatmap chart is sensitive to prisoners moving
between wings/prisons/PGD regions which can result in some "double counting".

To reduce this, we're now calculating the 'All' row for this chart using
data from the level above (e.g. for prison level, we sum values from
regional level which is not affected by prisoners moving wings).

* INC-626: fix tests

* INC-626: add footnotes to heatmap chart

* INC-626: Use `text-indent` CSS property for footnote list

Instead of more "esoteric" `padding-inline-start` CSS property - as
suggested in code review.

* INC-626: Heatmap footnote link style

Don't make it look like a link, as per prototype.

* INC-626: Fixed Heatmap chart footnotes indentation

* INC-626: Use correct hypen in Heatmap footnotes

Co-authored-by: Aldo "xoen" Giambelluca <aldo.giambelluca@digital.justice.gov.uk>