Sync `hmpps-template-typescript` and update dependencies (#302)
* Speeding up jest tests (hmpps-template-typescript#116)

This speeds up the running of jest tests by enabling isolatedModules
which has the effect of [disabling typechecking](https://kulshekhar.github.io/ts-jest/docs/getting-started/options/isolatedModules)

It also drastically reduces memory usage, allowing for running tests in
parallel locally at least.

On my laptop this reduces the time to run the tests in this project from
~14 seconds to ~4 seconds. On larger projects the effect is much more pronounced,
welcome-people-to-prison reduces build time from ~2mins, 20 seconds to ~25
seconds. In circle we still need to run in band but this is still
significantly faster than before, in WPIP it reduces the build by over 1 min.

Type checking is still available in the IDE, it is also part of the
husky pre-commit hook and run by circle as part of the build, so the
risk of type errors slipping through are very small.

(We could possibly add a typechecking stage before running jest and it
would still be much faster but not adding unless it becomes apparent
that we need it )

It would be worth to re-assess this after jest 29 as there seems to be
some fixes around a [memory issue](https://github.com/facebook/jest/issues/11956) that is part of node in versions
> 16.10

NB: This will not work if type declaration (`d.ts`) files contain enums
or any other constructs that generate javascript code. This seems to be
a bit of an anti-pattern anyway (see [here](https://lukasbehal.com/2017-05-22-enums-in-declaration-files/)).
Other tooling such as cypress will only allow you to import
types from these files.

* Fixing logging (hmpps-template-typescript#120)

There was an issue where we weren't sending trace info to app insights.
This is because appInsights needs to be imported before bunyan is
imported so it can do its instrumentation magic.

There was a related issue that obscured this. It was previously
impossible to test app insights locally as dotenv wasn't set up
correctly - it needed to happen before app insights is imported or app
insights would prevent the app starting up.

So this moves dotenv to dev dependencies and preloads it before running the app via start:dev.

This removes some code that is only relevant for local development.
It also means the application runs similar locally to how it would run
in docker or kubernetes - it just expects the environment variables to
be present.

Also moving the app insights import so it's very apparent that it's the first thing that happens when the app starts.

* Update dependencies 2022-09-28 (hmpps-template-typescript#121)

Co-authored-by: Andrew Lee <1517745+andrewrlee@users.noreply.github.com>
Co-authored-by: Paul Solecki <51918433+psoleckimoj@users.noreply.github.com>