INC-1269: Activate the previously-default-for-admissions level (#519)
…when there is no default-for-admissions incentive level in a prison but an inactive one exists. This is to allow the incentives-api to uphold it’s business rules.

If a prison had these levels:
• BAS active
• STD active
• ENH active and default
and is closed, then it ends up with:
• BAS inactive
• STD inactive
• ENH inactive and default

To activate the prison levels again, several api calls are needed – there is no single endpoint to "activate the default set of prison incentive levels".

If you were to try to make STD active and default, it seems that this should be possible, but the side effect is that ENH necessarily must be updated to non-default (there can only be 1 default per prison). Since ENH would remain inactive, this breaks the business rule that require BAS, STD & ENH to all be active. The api does not auto-activate required levels meaning that the client (this app) needs to carefully make update calls that have no side effects on other levels.

Therefore, if a prison is missing an active default level, the app must first check whether there is an inactive default level and activate that before attempting to activate STD.