Node.js 20
Node.js 20 is a few months old and it's becoming the next LTS (Long-term
support) release in a couple of weeks (2023-10-24).

https://nodejs.dev/en/about/releases/

We don't need to upgrade it now, but if the upgrade is uneventful it may
be a good idea to upgrade to it sooner rather than later.

Unit tests run fine.