Don't run tests before committing code
CI run tests. At best running tests is slowing you down but unfortunately
it's worse, some of the tests sometimes fail but only when run by
Husky (they are consistently fine when run manually or in CI).

This is particularly annoying when trying to doing something like an
interactive rebase and the failed test run prevents you from committing.