INC-1323: Upgrade to Node.js 20 (#627)
* Node.js 20

Node.js 20 is a few months old and it's becoming the next LTS (Long-term
support) release in a couple of weeks (2023-10-24).

https://nodejs.dev/en/about/releases/

We don't need to upgrade it now, but if the upgrade is uneventful it may
be a good idea to upgrade to it sooner rather than later.

Unit tests run fine.

* Don't run tests before committing code

CI run tests. At best running tests is slowing you down but unfortunately
it's worse, some of the tests sometimes fail but only when run by
Husky (they are consistently fine when run manually or in CI).

This is particularly annoying when trying to doing something like an
interactive rebase and the failed test run prevents you from committing.

* Don't specify `npm` version in `package.json`

The project doesn't do anything special with npm which requires a specific
version of it. Just the usual package management.

In practice use the most recent stable version of npm should be fine.

* Use a looser tag for the node base docker image

* README doesn't mention specific Node.js version

Added `nvm` command to install correct Node.js version.