Better handling of prisoners without any reviews
Although in theory all prisoners should have at
least an incentive review, in `dev` because of
misconfiguration (e.g. a prison without default
incentive level) or some of the testing (e.g. admitting
and instantly releasing a prisoner) some prisoner
have no incentive review records.

This causes [the API] to responds with a `404 Not Found`
which in turns means the UI can't be used to review
that prisoner (to quickly fix that data problem).

This change makes sure that this is handled more
gracefully and a minimal reviews page can be displayed
and the form to submit a new review work.

NOTE: If a prisoner doesn't exist at all the request
handlers wouldn't even reach the point of making
a request to Incentives API as the request to
Prisoner Search API would have already returned a
`404 Not Found`.

NOTE: The request handlers/templates were mostly
written in a way that could cope with missing data,
that's why this change is limited to the API client.

[the API]: https://incentives-api-dev.hmpps.service.justice.gov.uk/webjars/swagger-ui/index.html?configUrl=/v3/api-docs#/Maintain%20incentive%20reviews/getPrisonerIncentiveReviewHistory