INC-1399: Ability to use the UI to add an incentive review even for prisoner with no reviews (#963)
* Update templates to handle "no incentive reviews"

This is an edge case that mainly can happen in `dev`
as a result of manual testing and bad data but if
a prisoner has no current incentive level we now
fallback to "Not reviewed".

NOTE: The text 'Not reviewed' was chosen because of
some prior art in the reviews table:
https://github.com/ministryofjustice/hmpps-incentives-ui/blob/32bdb085dbaed0aac0ff65263ffd42b31068f887/server/views/pages/reviewsTable.njk#L88

* Better handling of prisoners without any reviews

Although in theory all prisoners should have at
least an incentive review, in `dev` because of
misconfiguration (e.g. a prison without default
incentive level) or some of the testing (e.g. admitting
and instantly releasing a prisoner) some prisoner
have no incentive review records.

This causes [the API] to responds with a `404 Not Found`
which in turns means the UI can't be used to review
that prisoner (to quickly fix that data problem).

This change makes sure that this is handled more
gracefully and a minimal reviews page can be displayed
and the form to submit a new review work.

NOTE: If a prisoner doesn't exist at all the request
handlers wouldn't even reach the point of making
a request to Incentives API as the request to
Prisoner Search API would have already returned a
`404 Not Found`.

NOTE: The request handlers/templates were mostly
written in a way that could cope with missing data,
that's why this change is limited to the API client.

[the API]: https://incentives-api-dev.hmpps.service.justice.gov.uk/webjars/swagger-ui/index.html?configUrl=/v3/api-docs#/Maintain%20incentive%20reviews/getPrisonerIncentiveReviewHistory

* Test that views render when prisoner has no reviews