Invalidate GHA cache
Just added a space in the `package-lock.json` (
a space in a JSON document shouldn't matter and this
file is often re-generated by tooling anyway).

I think the integration tests are currently failing
in CI because GHA is stubbornly using a cache
which was created before I updated the pipeline's
`node_build` job to run `npm run setup`.

The cached version was created before this tweak
which means GHA is not re-installing the dependencies
and **crucially** the Cypress binary - which is
installed via npm script (now controlled via
`npm run setup`/`hmpps-npm-script-run-allowlist`).

Or at least that's my theory.